Add OS includes
For some reason, I can't seem to add the "inc" directory to TMPx's
include path. I tried the short flag, I tried the long flag, I tried
adding the equal sign, I tried moving it to the end of the command,
but no matter what, it wouldn't let me write include paths starting
with 'os/'...

So as a hack I added a symlink to the 'os/' directory at the root.